Fixed overflow issue in Range(int, int)
As it was pointed out in issue #3  the int version of the Range method has a potential overflow issue.
Some casting was necessary to get the full int range supported.

+++ XorShift64.cs
@@ -87,7 +87,7 @@ namespace B83.Random.XorShift64
         }
         public int Range(int aMin, int aMax)
         {
-            return aMin + (int)Next() % (aMax - aMin);
+            return (int)((long)aMin + (uint)Next() % (uint)((long)aMax - (long)aMin));
         }
         public double Range(double aMin, double aMax)
         {
